Added Some Things. Main Updates: -refactoring of code -edited grass and trees, now grass is lighter and trees better realistic -organized a GUI files -now GameObjects are dynamic manage. If GameObject place on visible map then is updating and rendering -adited a format of map file. Now while world save or load as one file map -added sprites for Golem (idleLeft, idleRight, idleTop, idleBottom, runLeft, runRight, runTop, runBottom) -added sprites for Troll (idleLeft, idleRight, idleTop, idleBottom, runLeft, runRight, runTop, runBottom)
Building Editor Updates:
-added Menu Bar
-now load Building from Open Dialog Box (F6 or MainMenu->File->Load
Building)
-now save Building as "assets/building/testBuilding.txt" (F5 or
MainMenu->File->Save Building)
-now u can render statistics - colliders - of GameObjects using
MainMenu->Render options
-now u can selectr GameObjects

Map Editor Updates:
-added Menu Bar
-now the brush is more pleasant to use
-now save World Map as "world/world.txt" (F5 or MainMenu->File->Save
World)
-now load World Map from Open Dialog Box (F6 or MainMenu->File->Load
World)
-now u can render statistics of GameObjects using MainMenu->Render
options
-now datas are check the correct of Map format
-now you can add Water to tiles and it is animated

mush have

+++ headers/BrushSizes.hpp
@@ -68,13 +68,14 @@ std::vector<std::vector<std::vector<bool>>> brushes = {
 std::vector<std::vector<bool>> brush;
 
 
-void setBrushSize(int size) {
+void setBrushSize(short size) {
     brush = brushes[size];  // Przypisanie odpowiedniego pędzla
 }
 
 
 
 void coutBrush() {
+
     for (auto& row : brush) {
         for (bool cell : row)
             cout << cell << " ";
